Class renaming + Readme update
Gateway classes renamed to avoid confusion
pom updated accordingly
Readme updated
