+++ finish/gateway/src/main/java/io/openliberty/guides/gateway/GatewayJobResource.java
@@ -10,10 +10,11 @@
  *     IBM Corporation - Initial implementation
  *******************************************************************************/
 // end::copyright[]
-package io.openliberty.guides.gateway.client;
+package io.openliberty.guides.gateway;
 
 import java.util.concurrent.CompletionStage;
 
+import javax.inject.Inject;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
@@ -21,33 +22,48 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
+import org.eclipse.microprofile.rest.client.inject.RestClient;
 
+import io.openliberty.guides.gateway.client.JobClient;
+import io.openliberty.guides.models.JobList;
 import io.openliberty.guides.models.Job;
 import io.openliberty.guides.models.JobResult;
-import io.openliberty.guides.models.Jobs;
 
-@RegisterRestClient(baseUri = "http://job-service:9080")
 @Path("/jobs")
-public interface JobGatewayClient {
+public class GatewayJobResource {
 
-    // tag::getJobs[]
+    @Inject
+    @RestClient
+    private JobClient jobClient;
+
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public CompletionStage<Jobs> getJobs();
-    // end::getJobs[]
+    public CompletionStage<JobList> getJobs() {
+        return jobClient
+            .getJobs()
+            // tag::thenApplyAsync[]
+            .thenApplyAsync((jobs) -> {
+                return new JobList(jobs.getResults());
+            })
+            // end::thenApplyAsync[]
+            // tag::exceptionally[]
+            .exceptionally((ex) -> {
+                // Respond with empty list on error
+                return new JobList();
+            });
+            // end::exceptionally[]
+    }
 
-    // tag::getJob[]
     @GET
     @Path("{jobId}")
     @Produces(MediaType.APPLICATION_JSON)
-    public CompletionStage<JobResult> getJob(@PathParam("jobId") String jobId);
-    // end::getJob[]
+    public CompletionStage<JobResult> getJob(@PathParam("jobId") String jobId) {
+        return jobClient.getJob(jobId);
+    }
 
-    // tag::createJob[]
     @POST
     @Produces(MediaType.APPLICATION_JSON)
-    public CompletionStage<Job> createJob();
-    // end::createJob[]
-
+    public CompletionStage<Job> createJob() {
+        return jobClient.createJob();
+    }
 }
