+++ start/gateway/src/test/java/it/io/openliberty/guides/gateway/GatewayInventoryEndpointIT.java
@@ -31,7 +31,7 @@ import org.mockserver.junit.MockServerRule;
 import org.mockserver.model.HttpRequest;
 import org.mockserver.model.HttpResponse;
 
-public class InventoryGatewayEndpointTest {
+public class GatewayInventoryEndpointIT {
 
     private final String BASE_URL = "http://localhost:9080/api/systems";
