+++ start/inventory/src/test/java/it/io/openliberty/guides/inventory/InventoryEndpointIT.java
@@ -29,7 +29,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-public class InventoryEndpointTest {
+public class InventoryEndpointIT {
 
     private static final String port = System.getProperty("test.http.port");
     private static final String BASE_URL = "http://localhost:" + port + "/inventory/systems";
